Begin attacking `altScheduleGenerator`
