+++ scripts/package.json
@@ -9,9 +9,10 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@types/ical": "^0.8.0",
     "@types/node": "^16.11.7",
     "chalk": "^4.1.2",
-    "ical.js": "^1.4.0",
+    "ical": "^0.8.0",
     "node-fetch": "^3.1.0",
     "ts-node": "^10.4.0"
   }
